canvas_api v0.0.1 CanvasAPI.CanvasService

A service for viewing and manipulating canvases.

Functions

create(params, opts)

create(map, Keyword.t) ::
  {:ok, %CanvasAPI.Canvas{__meta__: term, blocks: term, creator: term, creator_id: term, id: term, inserted_at: term, is_template: term, native_version: term, slack_channel_ids: term, team: term, team_id: term, template: term, template_id: term, type: term, updated_at: term, version: term}} |
  {:error, Ecto.Changeset.t}

Create a new canvas from the given params.

The creator must provide a team and creator, and can optionally provide a template.

Options:

creator: %User{} (required) The user who created the canvas

team: %Team{} (required) The team to create the canvas in

template: map A map with "type" => "canvas" and an ID for the new canvas’s template. Ignored if the template is not found.

Examples

CanvasService.create(
  %{"is_template" => true},
  creator: current_user,
  team: current_team,
  template: %{"id" => "6ijSghOIflAjKVki5j0dpL", "type" => "canvas"})

delete(id, list)

delete(String.t, Keyword.t) ::
  {:ok, %CanvasAPI.Canvas{__meta__: term, blocks: term, creator: term, creator_id: term, id: term, inserted_at: term, is_template: term, native_version: term, slack_channel_ids: term, team: term, team_id: term, template: term, template_id: term, type: term, updated_at: term, version: term}} |
  nil |
  {:error, Ecto.Changeset.t}

Delete a canvas.

If the canvas is not found, returns nil. If the delete was invalid, returns {:error, changeset}. If it was successful, returns {:ok, canvas}.

Examples

CanvasService.delete(
  "6ijSghOIflAjKVki5j0dpL", team_id: "87ee9199-e2fa-49e6-9d99-16988af57fd5")

list(list)

list(Keyword.t) ::
  [%CanvasAPI.Canvas{__meta__: term, blocks: term, creator: term, creator_id: term, id: term, inserted_at: term, is_template: term, native_version: term, slack_channel_ids: term, team: term, team_id: term, template: term, template_id: term, type: term, updated_at: term, version: term}] |
  []

List canvases on behalf of a user.

Available filters:

user: %User{} (required) A user to list canvases for

only_templates: boolean List only templates canvases, including global templates, if they are defined.

Examples

CanvasService.list(user: current_user, only_templates: true)

show(id, list)

show(String.t, Keyword.t) ::
  %CanvasAPI.Canvas{__meta__: term, blocks: term, creator: term, creator_id: term, id: term, inserted_at: term, is_template: term, native_version: term, slack_channel_ids: term, team: term, team_id: term, template: term, template_id: term, type: term, updated_at: term, version: term} |
  nil

Show a canvas.

Examples

CanvasService.show(
  "6ijSghOIflAjKVki5j0dpL", team_id: "87ee9199-e2fa-49e6-9d99-16988af57fd5")

update(canvas, params, opts \\ [])

update(%CanvasAPI.Canvas{__meta__: term, blocks: term, creator: term, creator_id: term, id: term, inserted_at: term, is_template: term, native_version: term, slack_channel_ids: term, team: term, team_id: term, template: term, template_id: term, type: term, updated_at: term, version: term}, map, Keyword.t) ::
  {:ok, %CanvasAPI.Canvas{__meta__: term, blocks: term, creator: term, creator_id: term, id: term, inserted_at: term, is_template: term, native_version: term, slack_channel_ids: term, team: term, team_id: term, template: term, template_id: term, type: term, updated_at: term, version: term}} |
  {:error, Ecto.Changeset.t}

Update a canvas.

Examples

CanvasService.update(canvas, %{"is_template" => false})