Canvas API v0.0.1 CanvasAPI.CanvasTrackback

A “canvas trackback” is an event where a reference to a canvas was either added to or removed from another canvas.

Functions

add(target_id, source_id, acct_id)

add(String.t, String.t, String.t) ::
  %CanvasAPI.PulseEvent{__meta__: term, canvas: term, canvas_id: term, id: term, inserted_at: term, provider_name: term, provider_url: term, referencer: term, type: term, updated_at: term, url: term} |
  no_return

Create an “added” reference event, where the account user added a reference to the target canvas to the source canvas.

remove(target_id, source_id, acct_id)

remove(String.t, String.t, String.t) ::
  %CanvasAPI.PulseEvent{__meta__: term, canvas: term, canvas_id: term, id: term, inserted_at: term, provider_name: term, provider_url: term, referencer: term, type: term, updated_at: term, url: term} |
  no_return

Create a “removed” reference event, where the account user removed a reference to the target canvas from the source canvas.