Canvas API v0.0.1 CanvasAPI.JSONAPIView

A view of a resource according to the JSON API spec.

http://jsonapi.org

Functions

include(json_object, bool, resource, resource_func)

include(map, boolean, map, (() -> map | [map])) :: map

Include a resource object in a JSON object.

http://jsonapi.org/format/#fetching-includes

json_object(data, type \\ :data)

json_object(map, :data | :errors) :: map

Get a top-level JSON object of the data or errors type.

http://jsonapi.org/format/#document-top-level

link(resource_object, rel, url)

link(map, atom, String.t) :: map

Add a link to a JSON object or a resource object.

http://jsonapi.org/format/#document-links

paginate(json_object, params, page, func, action)

paginate(map, map, map, (... -> any), atom) :: map

Add pagination links to a JSON object.

http://jsonapi.org/format/#fetching-pagination

relate(resource_object, key, resource_identifier)

relate(map, atom, map) :: map

Add a relationship identity object to a resource object.

http://jsonapi.org/format/#document-resource-object-relationships

resource_object(resource, attributes)

resource_object(map, map) :: map

Build a resource object from a given resource and attributes.

http://jsonapi.org/format/#document-resource-objects

resource_object(resource, attributes, type)

resource_object(map, map, String.t) :: map