Canvas API v0.0.1 CanvasAPI.Router.Helpers

Module with named helpers generated from CanvasAPI.Router.

Functions

account_path(conn_or_endpoint, action)

account_path(conn_or_endpoint, action, params)

account_url(conn_or_endpoint, action)

account_url(conn_or_endpoint, action, params)

bulk_path(conn_or_endpoint, action)

bulk_path(conn_or_endpoint, action, params)

bulk_url(conn_or_endpoint, action)

bulk_url(conn_or_endpoint, action, params)

callback_path(conn_or_endpoint, action)

callback_path(conn_or_endpoint, action, params)

callback_url(conn_or_endpoint, action)

callback_url(conn_or_endpoint, action, params)

meta_path(conn_or_endpoint, action)

meta_path(conn_or_endpoint, action, params)

meta_url(conn_or_endpoint, action)

meta_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

slack_path(conn_or_endpoint, action)

slack_path(conn_or_endpoint, action, params)

slack_url(conn_or_endpoint, action)

slack_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

team_canvas_path(conn_or_endpoint, action, team_id)

team_canvas_path(conn_or_endpoint, action, team_id, params)

team_canvas_path(conn_or_endpoint, action, team_id, id, params)

team_canvas_pulse_event_path(conn_or_endpoint, action, team_id, canvas_id)

team_canvas_pulse_event_path(conn_or_endpoint, action, team_id, canvas_id, params)

team_canvas_pulse_event_url(conn_or_endpoint, action, team_id, canvas_id)

team_canvas_pulse_event_url(conn_or_endpoint, action, team_id, canvas_id, params)

team_canvas_url(conn_or_endpoint, action, team_id)

team_canvas_url(conn_or_endpoint, action, team_id, params)

team_canvas_url(conn_or_endpoint, action, team_id, id, params)

team_channel_path(conn_or_endpoint, action, team_id)

team_channel_path(conn_or_endpoint, action, team_id, params)

team_channel_url(conn_or_endpoint, action, team_id)

team_channel_url(conn_or_endpoint, action, team_id, params)

team_path(conn_or_endpoint, action)

team_path(conn_or_endpoint, action, params)

team_path(conn_or_endpoint, action, id, params)

team_template_path(conn_or_endpoint, action, team_id)

team_template_path(conn_or_endpoint, action, team_id, params)

team_template_url(conn_or_endpoint, action, team_id)

team_template_url(conn_or_endpoint, action, team_id, params)

team_url(conn_or_endpoint, action)

team_url(conn_or_endpoint, action, params)

team_url(conn_or_endpoint, action, id, params)

team_user_path(conn_or_endpoint, action, team_id)

team_user_path(conn_or_endpoint, action, team_id, params)

team_user_url(conn_or_endpoint, action, team_id)

team_user_url(conn_or_endpoint, action, team_id, params)

unfurl_path(conn_or_endpoint, action)

unfurl_path(conn_or_endpoint, action, params)

unfurl_url(conn_or_endpoint, action)

unfurl_url(conn_or_endpoint, action, params)

upload_signature_path(conn_or_endpoint, action)

upload_signature_path(conn_or_endpoint, action, params)

upload_signature_url(conn_or_endpoint, action)

upload_signature_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.