Canvas API v0.0.1 CanvasAPI.SlackParser

Provides functionality for parsing Slack messages.

Functions

to_text(message, result \\ "", state \\ [state: :garbage])

to_text(String.t, String.t, Keyword.t) :: String.t

Convert a Slack message to plain text by stripping brackets and leaving only link text and labels.

Examples

iex> CanvasAPI.SlackParser.to_text("Foo <bar>")
"Foo bar"

iex> CanvasAPI.SlackParser.to_text("Foo <!here|@here> Bar")
"Foo @here Bar"

iex> CanvasAPI.SlackParser.to_text("Foo <https://www.example.com>")
"Foo https://www.example.com"