canvas_api v0.0.1 CanvasAPI.Web

A module that keeps using definitions for controllers, views and so on.

This can be used in your application as:

use CanvasAPI.Web, :controller
use CanvasAPI.Web, :view

The definitions below will be executed for every view, controller, etc, so keep them short and clean, focused on imports, uses and aliases.

Do NOT define functions inside the quoted expressions below.

Functions

channel()

controller()

model()

router()

service()

view()

Macros

__using__(which)

When used, dispatch to the appropriate controller/view/etc.